3AgentGCExec/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.30.4
packaging==24.0
pyautogen==0.2.27
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
